Consolidated TournamentEndpointTests
The TournamentEndpointTests now only use one database.
